+++ releasenotes.md
@@ -33,6 +33,7 @@
 - Added missing `@clone_aligned` and add checks to `@tclone`
 - Comparing a distinct type with an enum with an inline distinct type failed unexpectedly.
 - The `%s` would not properly print function pointers.
+- Compiler crash when passing an untyped list as an argument to `assert` #2108.
 
 ### Stdlib changes
 - Hash functions for integer vectors and arrays.